Add grafting API function to get graftable augs
